@@ -1,6 +1,4 @@
-'use strict';
-
-exports = module.exports = class Semaphore {
+export default class Semaphore {
 
 	constructor() {
 		this._awaiters = [];
